fix a typo in the elementwise pow_ kernel function

+++ src/cl_programs/elementwise.rs
@@ -102,7 +102,7 @@ where
         }}
 
         inline {ltype} pow_(const {ltype} left, const {rtype} right) {{
-            return pow(left, right);
+            return powf(left, right);
         }}
 
         inline {ltype} sub(const {ltype} left, const {rtype} right) {{
